% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqdatapipeline-package.R
\docType{package}
\name{aqdatapipeline-package}
\alias{aqdatapipeline}
\alias{aqdatapipeline-package}
\title{aqdatapipeline: Download and process air quality data}
\description{
The package downloads AURN data to a local partitioned Parquet dataset and
calculates daily, monthly, and annual means from hourly observations.
}
\author{
\strong{Maintainer}: Joseph Grealy \email{joseph@example.com}

}
\keyword{internal}
