% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_period_means.R
\name{calculate_period_means}
\alias{calculate_period_means}
\title{Calculate period means from hourly observations}
\usage{
calculate_period_means(hourly_data, averaging)
}
\arguments{
\item{hourly_data}{Hourly AURN observations.}

\item{averaging}{One of \code{Daily}, \code{Monthly}, or \code{Annual}.}
}
\value{
A normalised tibble of calculated means.
}
\description{
Calculate period means from hourly observations
}
