% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_id.R
\name{chunk_id}
\alias{chunk_id}
\title{Create an exact download chunk identifier}
\usage{
chunk_id(data_type, pollutant, years)
}
\arguments{
\item{data_type}{Download data type.}

\item{pollutant}{Openair pollutant identifier.}

\item{years}{Integer vector of years in the chunk.}
}
\value{
A stable character identifier.
}
\description{
Create an exact download chunk identifier
}
