% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_chunk.R
\name{download_chunk}
\alias{download_chunk}
\title{Download one AURN data chunk}
\usage{
download_chunk(
  sitecodes,
  years,
  pollutant,
  data_type,
  import_fun = openair::importAURN
)
}
\arguments{
\item{sitecodes}{Lower-case AURN site codes.}

\item{years}{Integer vector of years.}

\item{pollutant}{Openair pollutant identifier.}

\item{data_type}{Download data type.}

\item{import_fun}{Function compatible with \code{\link[openair:importUKAQ-wrapper]{openair::importAURN()}}.}
}
\value{
A list with \code{status}, \code{data}, and \code{messages}.
}
\description{
Download one AURN data chunk
}
