% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_schema.R
\name{normalise_schema}
\alias{normalise_schema}
\title{Normalise data to the AURN database schema}
\usage{
normalise_schema(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A tibble using the standard database schema.
}
\description{
Renames known openair columns, removes non-database columns, adds typed
missing columns, and returns columns in their standard order.
}
