% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_connection_ids.R
\name{open_connection_ids}
\alias{open_connection_ids}
\title{Get identifiers for all open R connections}
\usage{
open_connection_ids()
}
\value{
A character vector of open connection identifiers.
}
\description{
Get identifiers for all open R connections
}
\keyword{internal}
