% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_columns.R
\name{partition_columns}
\alias{partition_columns}
\title{AURN Parquet partition columns}
\usage{
partition_columns()
}
\value{
A character vector of partition column names.
}
\description{
AURN Parquet partition columns
}
\keyword{internal}
