% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_status_log.R
\name{read_status_log}
\alias{read_status_log}
\title{Read the structured chunk status log}
\usage{
read_status_log(path)
}
\arguments{
\item{path}{Status CSV path.}
}
\value{
A tibble.
}
\description{
Read the structured chunk status log
}
