% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_columns.R
\name{standard_columns}
\alias{standard_columns}
\title{Standard AURN database columns}
\usage{
standard_columns()
}
\value{
A character vector containing the ordered database schema.
}
\description{
Standard AURN database columns
}
\keyword{internal}
