% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config.R
\name{validate_config}
\alias{validate_config}
\title{Validate pipeline configuration}
\usage{
validate_config(config)
}
\arguments{
\item{config}{A configuration list created by \code{\link[=default_config]{default_config()}}.}
}
\value{
\code{config}, invisibly.
}
\description{
Validate pipeline configuration
}
